fixed code motion-framer, logo, etc
