+++ src/components/landing/hero.tsx
@@ -1,11 +1,15 @@
 "use client"
 
-import { motion } from "framer-motion"
+import { HTMLAttributes } from "react"
+import { motion, MotionProps } from "framer-motion"
 import { useRouter } from "next/navigation"
 import { Button } from "@/components/ui/button"
 import { useLanguage } from "@/contexts/LanguageContext"
 import en from "@/translations/en.json"
 import uk from "@/translations/uk.json"
+
+// type for motion.div
+type MotionDivProps = MotionProps & HTMLAttributes<HTMLDivElement>
 
 export default function Hero() {
   const router = useRouter()
@@ -21,6 +25,7 @@ export default function Hero() {
             initial={{ opacity: 0, y: 20 }}
             animate={{ opacity: 1, y: 0 }}
             transition={{ duration: 0.8 }}
+            {...({} as MotionDivProps)}
           >
             <h1 className="text-3xl md:text-4xl lg:text-6xl font-bold leading-tight bg-clip-text text-transparent bg-gradient-to-r from-purple-600 to-purple-800 dark:from-blue-400 dark:to-purple-600 filter drop-shadow-lg py-2">
               {t.hero.title}
@@ -33,6 +38,7 @@ export default function Hero() {
               initial={{ opacity: 0, y: 20 }}
               animate={{ opacity: 1, y: 0 }}
               transition={{ delay: 0.2, duration: 0.5 }}
+              {...({} as MotionDivProps)}
             >
               <Button
                 size="lg"
