+++ src/translations/uk.json
@@ -10,43 +10,37 @@
     "learnMore": "Дізнатися більше"
   },
   "features": {
-    "title": "Ключові особливості",
+    "title": "Відкрийте для себе наші потужні функції",
     "smartCalendar": {
       "title": "Розумний календар",
       "description": "Легко плануйте свій день за допомогою нашої інтелектуальної інтеграції календаря.",
-      "details": "Наш Розумний Календар використовує ШІ для вивчення ваших шаблонів продуктивності та планування завдань у той час, коли ви найімовірніше виконаєте їх ефективно.",
-      "usageExample": "// Запланувати завдання з Розумним Календарем\nsmartCalendar.scheduleTask({\n  title: 'Завершити проектну пропозицію',\n  duration: '2 години',\n  priority: 'високий'\n});"
+      "details": "Наш Розумний Календар використовує ШІ для вивчення ваших шаблонів продуктивності та планування завдань у той час, коли ви найімовірніше виконаєте їх ефективно."
     },
     "taskManagement": {
       "title": "Управління завданнями",
       "description": "Створюйте, організовуйте та пріоритезуйте завдання за допомогою нашого інтуїтивного інтерфейсу.",
-      "details": "Наша система управління завданнями дозволяє розбивати складні проекти на керовані підзавдання, встановлювати повторювані завдання та використовувати користувацькі мітки для категоризації вашої роботи.",
-      "usageExample": "// Створити нове завдання з підзавданнями\nconst projectTask = taskManager.createTask({\n  title: 'Запуск нового продукту',\n  dueDate: '2023-12-31',\n  subtasks: [\n    { title: 'Завершити дизайн', assignee: 'Аліса' },\n    { title: 'Написати документацію', assignee: 'Богдан' }\n  ]\n});"
-    },
+      "details": "Наша система управління завданнями дозволяє розбивати складні проекти на керовані підзавдання, встановлювати повторювані завдання та використовувати користувацькі мітки для категоризації вашої роботи."
+      },
     "aiAssistant": {
       "title": "ШІ-асистент",
       "description": "Отримуйте розумні пропозиції та автоматизуйте управління завданнями.",
-      "details": "Наш ШІ-асистент аналізує ваші шаблони виконання завдань та дані про продуктивність, щоб надати інформацію про те, коли ви найбільш ефективні.",
-      "usageExample": "// Отримати пропозиції завдань на основі ШІ\nconst suggestions = await aiAssistant.getSuggestions({\n  user: currentUser,\n  timeFrame: 'сьогодні'\n});\n\nsuggestions.forEach(suggestion => {\n  console.log(`Пропозиція: ${suggestion.text}`);\n});"
-    },
+      "details": "Наш ШІ-асистент аналізує ваші шаблони виконання завдань та дані про продуктивність, щоб надати інформацію про те, коли ви найбільш ефективні."
+     },
     "teamCollaboration": {
       "title": "Командна співпраця",
       "description": "Ефективно працюйте разом з обміном завданнями в реальному часі.",
-      "details": "Наші функції командної співпраці включають спільні дошки проектів, оновлення статусу завдань у реальному часі та обмін повідомленнями в додатку.",
-      "usageExample": "// Поділитися завданням з членами команди\nconst sharedTask = await taskManager.shareTask({\n  taskId: '123456',\n  sharedWith: ['team@example.com'],\n  permissions: ['перегляд', 'редагування']\n});\n\nconsole.log(`Завдання поділено з ${sharedTask.sharedWith.length} членами команди`);"
-    },
+      "details": "Наші функції командної співпраці включають спільні дошки проектів, оновлення статусу завдань у реальному часі та обмін повідомленнями в додатку."
+     },
     "quickActions": {
       "title": "Швидкі дії",
       "description": "Підвищуйте продуктивність за допомогою гарячих клавіш та швидкого створення завдань.",
-      "details": "Швидкі дії дозволяють додавати завдання з будь-якого місця в додатку за допомогою налаштовуваних комбінацій клавіш та голосових команд.",
-      "usageExample": "// Зареєструвати користувацьку швидку дію\nquickActions.register({\n  name: 'Створити термінове завдання',\n  shortcut: 'Ctrl+Shift+U',\n  action: () => {\n    taskManager.createTask({ title: 'Термінове завдання', priority: 'високий' });\n  }\n});"
-    },
+      "details": "Швидкі дії дозволяють додавати завдання з будь-якого місця в додатку за допомогою налаштовуваних комбінацій клавіш та голосових команд."
+      },
     "smartNotifications": {
       "title": "Розумні сповіщення",
       "description": "Ніколи не пропускайте важливі завдання завдяки інтелектуальним нагадуванням.",
-      "details": "Наша система розумних сповіщень використовує машинне навчання для визначення найкращого часу для нагадування про завдання, враховуючи такі фактори, як ваше місцезнаходження та минула поведінка.",
-      "usageExample": "// Налаштувати розумне сповіщення\nsmartNotifications.create({\n  taskId: '789012',\n  message: 'Не забудьте підготуватися до завтрашньої зустрічі',\n  smartTiming: true,\n  channels: ['push', 'email']\n});"
-    }
+      "details": "Наша система розумних сповіщень використовує машинне навчання для визначення найкращого часу для нагадування про завдання, враховуючи такі фактори, як ваше місцезнаходження та минула поведінка."
+     }
   },
   "footer": {
     "terms": "Умови",
@@ -99,26 +93,7 @@
     }
   },
   "featuresPage": {
-    "title": "Відкрийте для себе наші потужні функції",
-    "subtitle": "Plan My Day пропонує комплексний набір інструментів для підвищення вашої продуктивності та оптимізації робочого процесу.",
-    "getStarted": "Почати",
-    "analytics": {
-      "title": "Аналітична панель",
-      "description": "Отримуйте уявлення про вашу продуктивність за допомогою детальної аналітики."
-    },
-    "crossPlatform": {
-      "title": "Кросплатформенна синхронізація",
-      "description": "Отримуйте доступ до своїх завдань звідусіль, на будь-якому пристрої."
-    },
-    "dataSecurity": {
-      "title": "Безпека даних",
-      "description": "Ваші дані зашифровані та надійно зберігаються."
-    },
-    "mobileApp": {
-      "title": "Мобільний додаток",
-      "description": "Залишайтеся організованими в дорозі за допомогою нашого мобільного додатку."
-    },
-    "startPlanning": "Почніть планувати свій день сьогодні!"
+    "subtitle": "Plan My Day пропонує комплексний набір інструментів для підвищення вашої продуктивності та оптимізації робочого процесу."
   },
   "legal": {
     "terms": {
